+++ README.md
@@ -16,22 +16,23 @@ THREADS_AUTH_CODE_REDIRECT_URI=<REDIRECT_URL>
 ```
 
 ```rs
-use rusty_meta_threads;
+use rusty_meta_threads::{auth, profiles};
 
 //------------------------------
-// client-side
-let login_url = rusty_meta_threads::get_threads_login_url();
-// now we can redirect the end-user to `login_url`
+// CLIENT-SIDE
+
+let login_url = auth::get_threads_login_url();
+// now we can redirect the end-user to `login_url`,
 // once the end-user logs in successfully, Threads will send
-// a request to <REDIRECT_URL>, which we'll handle
+// a request to `<REDIRECT_URL>?code=<CODE>`, which we'll handle
 // on the server-side
 
 //------------------------------
-// server-side
+// SERVER-SIDE
 
-// inside an HTTP endpoint handler logic
+// inside the HTTP endpoint handler logic for e.g. my-app.domain/auth
 let code =
-    rusty_meta_threads::get_code_from_redirect_uri("<REDIRECT_URL>");
+    auth::get_code_from_redirect_uri("<REDIRECT_URL_WITH_CODE>");
 
 // `code` can now be exchanged for a short-lived bearer token
 let short_lived_token =
@@ -43,7 +44,7 @@ let short_lived_token =
 // `short_lived_token` can be exchanged for
 // a long-lived bearer token
 let long_lived_token =
-    rusty_meta_threads::get_long_lived_bearer_token(&short_lived_token)
+    auth::get_long_lived_bearer_token(&short_lived_token)
         .await
         .unwrap()
         .access_token;
@@ -51,14 +52,14 @@ let long_lived_token =
 // either `short_lived_token` or `long_lived_token` can be used
 // to make API calls against Threads, for example:
 let profile_info =
-    rusty_meta_threads::get_profile_info(&long_lived_token)
+    profiles::get_profile_info(&long_lived_token)
         .await
         .unwrap();
 
 // long-lived token can be refreshed
 // as long as it doesn't expire yet
 let refreshed_token =
-    rusty_meta_threads::refresh_long_lived_bearer_token(&long_lived_token)
+    auth::refresh_long_lived_bearer_token(&long_lived_token)
         .await
         .unwrap()
         .access_token;
